don't error out if some of the dlls needed by plink are missing - happens with 32-bit builds
git-svn-id: https://xpra.org/svn/Xpra/tags/v3.0.x@27083 3bb7dfac-3a0b-4e04-842a-767bc560f471

+++ src/win32/MINGW_BUILD.sh
@@ -358,7 +358,9 @@ if [ "${BUNDLE_PUTTY}" == "1" ]; then
 	fi
 	cp -fn "${TORTOISESVN}/bin/TortoisePlink.exe" "${DIST}/Plink.exe"
 	for dll in vcruntime140.dll msvcp140.dll vcruntime140_1.dll; do
-		cp "/c/Windows/System32/$dll" "${DIST}/"
+		if [ -e "/c/Windows/System32/$dll" ]; then
+			cp "/c/Windows/System32/$dll" "${DIST}/"
+		fi
 	done
 	#are we meant to include those DLLs?
 	#rsync -rplogt "${TORTOISESVN}/bin/"*dll "${DIST}/"
